refactor(review-step3): Money 타입 변경 int => long, 메서드명 변경 changeUnit => exchangeLottoPurchasableCount
* BigInteger 같은 타입으로 변경할필요까지는 없어보임.

+++ src/main/java/step3/domain/LottoShop.java
@@ -21,7 +21,7 @@ public class LottoShop {
         this.income.earn(money);
     }
 
-    public int countOfSelling() {
+    public long countOfSelling() {
         return this.income.get();
     }
 }
